Добавлено логирование, подправлены entity модели, подправлены контроллеры авторизации

+++ protocol-app/src/main/java/com/execute/protocol/app/models/GameInfo.java
@@ -8,6 +8,6 @@ import lombok.*;
 @AllArgsConstructor
 @NoArgsConstructor
 public class GameInfo {
-    private long answer;
-    private long event;
+    private int answer;
+    private int event;
 }
