+++ protocol-core/src/main/java/com/execute/protocol/core/repositories/TargetRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TargetRepository extends JpaRepository<Target, Long> {
+public interface TargetRepository extends JpaRepository<Target, Integer> {
 }
